sistema_clinica/requirements.txt
Flask>=2.0
pymysql
Werkzeug

